requirements.txt
openai==0.25.0
beautifulsoup4==4.11.1
gunicorn==20.1.0
flask==2.2.2
requests==2.28.1
google-cloud-translate==3.9.0
python-dotenv==0.21.0